Adding release target to makefile
This automates the creation of a new tagged release.

Run
    make release VERSION=0.9.8
to create an annotated tag object "0.9.8" on the master branch. Afterwards, master
will be merged back in to develop, in order to make the new tag visible there, too.

This target will refuse to run if the git working directory contains uncommitted
changes. Together with the dependency on the "all" target, this will enforce
an up to date production version of the JS.