Adding reconnect handler for offline resilience
This makes the app robust against dropped connections or short server
downtimes. After reconnectiing, the client re-joins the group and broadcasts
the local state.

Note: There is some redundancy in the case of multiple clients reconnecting
at the same time. Each client will broadcast the local state and will then
broadcast it again because another client asks "who is there". That seems
acceptable, though, as a poker round is not expected to have more than
a handful of users.