Revert "Add a manifest file for offline usage"
This reverts commit d1f50e68057b085ef90bf49568a8c7297411a8dd.

The offline caching caused socket.io connectivity issues in Firefox.
When restarting the server, Firefox would not reconnect the websocket,
the console.log debug output claims that requests to the server are made,
but the network tab and the server log show nothing.

Automatic reconnect after a network interruption has more value than
caching a couple of KB through an offline manifest.