Use Knockout debug version in production, too
The monkey patching for subscriptions

    isDisposed = true

that is used for pausing subscription handling, does not work with the
pre-minified version of Knockout, because the private variable is not
named isDisposed anymore.

When minifying the debug version through closure compiler, the patching
works as expected.

There is a bandwidth penalty of 20KB for this.